Move S3Exception to "s3mock" package, it's only used here.
