+++ server/src/main/java/com/adobe/testing/s3mock/S3Exception.java
@@ -14,12 +14,14 @@
  *  limitations under the License.
  */
 
-package com.adobe.testing.s3mock.store;
+package com.adobe.testing.s3mock;
+
+import com.adobe.testing.s3mock.dto.ErrorResponse;
 
 /**
- * {@link RuntimeException} to communicate general S3 errors. These are handled by {@code
- * com.adobe.testing.s3mock.util.S3ExceptionResolver}, mapped to {@link
- * com.adobe.testing.s3mock.dto.ErrorResponse} and serialized.
+ * {@link RuntimeException} to communicate general S3 errors.
+ * These are handled by {@link S3MockConfiguration.S3MockExceptionHandler},
+ * mapped to {@link ErrorResponse} and serialized.
  */
 public class S3Exception extends RuntimeException {
 
@@ -28,8 +30,7 @@ public class S3Exception extends RuntimeException {
   private final String message;
 
   /**
-   * Creates a new S3Exception to be mapped as an
-   * {@link com.adobe.testing.s3mock.dto.ErrorResponse}.
+   * Creates a new S3Exception to be mapped as an {@link ErrorResponse}.
    *
    * @param status The Error Status.
    * @param code The Error Code.
